+++ hypr/packages/system.sh
@@ -7,7 +7,7 @@ _PKGDIR="$(cd "$(dirname "$0")" && pwd)"
 case "$HYPRTK_PM" in
 pacman)
     PKGS=(sddm blueman pacman-contrib fzf font-manager awesome-terminal-fonts
-          ttf-font-awesome ttf-fira-sans ttf-fira-code ttf-firacode-nerd exa
+          otf-font-awesome ttf-fira-sans ttf-fira-code ttf-firacode-nerd eza
           python-pip python-psutil python-rich python-click xdg-desktop-portal-gtk
           xdg-user-dirs xdg-user-dirs-gtk os-prober polkit-gnome gnome-keyring pcp
           pcp-gui gtk4-layer-shell hyprpicker)
@@ -21,7 +21,7 @@ apt)
           policykit-1-gnome gnome-keyring libgtk4-layer-shell0 hyprpicker)
     ;;
 dnf)
-    PKGS=(sddm blueman fzf font-manager fontawesome-fonts fira-code-fonts eza
+    PKGS=(sddm blueman fzf font-manager fontawesome-fonts-all fira-code-fonts eza
           python3-pip python3-psutil python3-rich python3-click xdg-desktop-portal-gtk
           xdg-user-dirs xdg-user-dirs-gtk os-prober polkit-gnome gnome-keyring
           gtk4-layer-shell hyprpicker)
@@ -29,17 +29,17 @@ dnf)
 zypper)
     PKGS=(sddm blueman fzf font-manager fontawesome-fonts fira-code-fonts eza
           python3-pip python3-psutil python3-rich python3-click xdg-desktop-portal-gtk
-          xdg-user-dirs xdg-user-dirs-gtk os-prober polkit-gnome gnome-keyring
-          gtk4-layer-shell hyprpicker)
+          xdg-user-dirs           xdg-user-dirs-gtk os-prober polkit-gnome gnome-keyring
+          libgtk4-layer-shell0 hyprpicker)
     ;;
 xbps)
-    PKGS=(sddm blueman fzf font-manager font-awesome fira-code eza python3-pip
+    PKGS=(sddm blueman fzf fontmanager font-awesome font-firacode eza python3-pip
           python3-psutil python3-rich python3-click xdg-desktop-portal-gtk
           xdg-user-dirs xdg-user-dirs-gtk os-prober polkit-gnome gnome-keyring
           gtk4-layer-shell hyprpicker)
     ;;
 apk)
-    PKGS=(sddm blueman fzf font-manager font-awesome fira-code eza py3-pip
+    PKGS=(sddm blueman fzf font-manager font-awesome font-fira-code-nerd eza py3-pip
           py3-psutil py3-rich py3-click xdg-desktop-portal-gtk xdg-user-dirs
           xdg-user-dirs-gtk os-prober polkit-gnome gnome-keyring gtk4-layer-shell
           hyprpicker)