+++ installer/hyprtk-bar/install.sh
@@ -110,7 +110,7 @@ declare -A DEPS
 DEPS[pacman]="gtk3 gtk-layer-shell gdk-pixbuf2 pango cairo gobject-introspection-runtime python python-gobject python-cairo python-pip"
 DEPS[apt]="gir1.2-gtk-3.0 gir1.2-gtklayershell-0.1 gir1.2-gdkpixbuf-2.0 gir1.2-pango-1.0 gir1.2-cairo-1.0 gir1.2-xlib-2.0 python3-gi python3-gi-cairo python3-venv python3-pip"
 DEPS[dnf]="gtk3 gtk-layer-shell gdk-pixbuf2 pango cairo gobject-introspection python3-gobject python3-cairo python3-pip"
-DEPS[zypper]="typelib-1_0-Gtk-3_0 typelib-1_0-GtkLayerShell-0_1 typelib-1_0-GdkPixbuf-2_0 typelib-1_0-Pango-1_0 typelib-1_0-cairo-1_0 typelib-1_0-xlib-2_0 python3-gobject python3-gobject-Gdk python3-gobject-cairo python3-pip"
+DEPS[zypper]="typelib-1_0-Gtk-3_0 typelib-1_0-GtkLayerShell-0_1 typelib-1_0-GdkPixbuf-2_0 typelib-1_0-Pango-1_0 girepository-1_0 python3-gobject python3-gobject-Gdk python3-gobject-cairo python3-pip"
 # Void/Alpine ship the GI typelibs in the -devel/-dev subpackages, not the base
 # lib packages.
 DEPS[xbps]="gtk+3-devel gtk-layer-shell-devel gdk-pixbuf-devel pango-devel cairo-devel gobject-introspection python3-gobject python3-cairo python3-pip"